reorganizing files, renamed "Replacement Weapons" to "Vanilla Weapons", changed version number to 2.0.0
