+++ ULTRAKIT.Data/ULTRAKIT.Data.csproj
@@ -112,7 +112,7 @@
     <Compile Include="ScriptableObjects\Arm.cs" />
     <Compile Include="ScriptableObjects\AssetBundleData.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ScriptableObjects\HatRegistry.cs" />
+    <Compile Include="ScriptableObjects\Hats.cs" />
     <Compile Include="ScriptableObjects\UKSpawnable.cs" />
     <Compile Include="ScriptableObjects\Weapon.cs" />
   </ItemGroup>
