+++ ULTRAKIT.Extensions/Managers/HatsManager.cs
