+++ ULTRAKIT.Extensions/UKClassExtensions/GunSetter.cs
