+++ ULTRAKIT.Loader/Loaders/HatLoader.cs
