+++ ULTRAKIT.Loader/Patches/SpawnerArmPatch.cs
