+++ ULTRAKIT/ULTRAKIT.csproj
@@ -112,8 +112,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Cheats_Commands\ActivateHats.cs" />
-    <Compile Include="Cheats_Commands\AltSetter.cs" />
+    <Compile Include="Cheats\ActivateHats.cs" />
+    <Compile Include="Commands\AltSetter.cs" />
     <Compile Include="Initializer.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -122,7 +122,7 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <Compile Include="Cheats_Commands\Refresher.cs" />
+    <Compile Include="Cheats\Refresher.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ULTRAKIT.Data\ULTRAKIT.Data.csproj">
